refactor(portfolio): reorganize asset sections
