+++ apps/ledger-live-mobile/src/mvvm/features/Portfolio/components/PortfolioEmptySection/PortfolioNoAccountsContent.tsx
@@ -4,8 +4,8 @@ import { QuickActionsCtas, TransferDrawer } from "LLM/features/QuickActions";
 import MarketBanner from "LLM/features/MarketBanner";
 import { ScreenName } from "~/const";
 import { PortfolioBannersSection } from "../PortfolioBannersSection";
-import { PortfolioCryptosSection } from "../PortfolioCryptosSection";
-import { PortfolioStablecoinsSection } from "../PortfolioStablecoinsSection";
+import { PortfolioCryptosSection } from "LLM/features/WalletAssets/components/CryptosSection";
+import { PortfolioStablecoinsSection } from "LLM/features/WalletAssets/components/StablecoinsSection";
 import AddAccountDrawer from "LLM/features/Accounts/screens/AddAccount";
 import { useTranslation } from "~/context/Locale";
 import TrackScreen from "~/analytics/TrackScreen";
