+++ apps/ledger-live-mobile/src/mvvm/features/Portfolio/components/PortfolioEmptySection/PortfolioNoSignerContent.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Box } from "@ledgerhq/lumen-ui-rnative";
 import { QuickActionsCtas, TransferDrawer } from "LLM/features/QuickActions";
 import { ScreenName, NavigatorName } from "~/const";
-import { PortfolioCryptosSection } from "../PortfolioCryptosSection";
-import { PortfolioStablecoinsSection } from "../PortfolioStablecoinsSection";
+import { PortfolioCryptosSection } from "LLM/features/WalletAssets/components/CryptosSection";
+import { PortfolioStablecoinsSection } from "LLM/features/WalletAssets/components/StablecoinsSection";
 import { PortfolioBannersSection } from "../PortfolioBannersSection";
 import MarketBanner from "LLM/features/MarketBanner";
 import TrackScreen from "~/analytics/TrackScreen";
