+++ apps/ledger-live-mobile/src/mvvm/features/Portfolio/components/index.ts
@@ -2,10 +2,10 @@ export { PortfolioAllocationsSection } from "./PortfolioAllocationsSection";
 export { PortfolioAssetsSection } from "./PortfolioAssetsSection";
 export { PortfolioBalanceSection } from "./PortfolioBalanceSection";
 export { PortfolioCarouselSection } from "./PortfolioCarouselSection";
-export { PortfolioCategorizedAssetsSection } from "./PortfolioCategorizedAssetsSection";
-export { PortfolioCryptosSection } from "./PortfolioCryptosSection";
-export { SeeAllAssetsButton } from "./SeeAllAssetsButton";
-export { PortfolioStablecoinsSection } from "./PortfolioStablecoinsSection";
+export { PortfolioCategorizedAssetsSection } from "LLM/features/WalletAssets/views/CategorizedAssetsSection";
+export { PortfolioCryptosSection } from "LLM/features/WalletAssets/components/CryptosSection";
+export { SeeAllAssetsButton } from "LLM/features/WalletAssets/components/SeeAllAssetsButton";
+export { PortfolioStablecoinsSection } from "LLM/features/WalletAssets/components/StablecoinsSection";
 export { PortfolioEmptySection } from "./PortfolioEmptySection";
 export { PortfolioHeaderSection } from "./PortfolioHeaderSection";
 export { PortfolioOperationsSection } from "./PortfolioOperationsSection";
