+++ apps/ledger-live-mobile/src/mvvm/features/WalletAssets/components/CryptosSection/__tests__/AssetListItem.test.tsx
