+++ apps/ledger-live-mobile/src/mvvm/features/WalletAssets/components/CryptosSection/index.tsx
