+++ apps/ledger-live-mobile/src/mvvm/features/WalletAssets/components/StablecoinsSection/__tests__/shared.ts
