+++ apps/ledger-live-mobile/src/mvvm/features/WalletAssets/components/StablecoinsSection/__tests__/usePortfolioStablecoinsSectionViewModel.test.ts
