+++ apps/ledger-live-mobile/src/mvvm/features/WalletAssets/components/StablecoinsSection/usePortfolioStablecoinsSectionViewModel.ts
