+++ apps/ledger-live-mobile/src/mvvm/features/WalletAssets/views/CategorizedAssetsSection/index.tsx
